refactor: add comment and move inittenancy to leader file
