+++ agent/consul/leader_ce_test.go
@@ -5,7 +5,19 @@
 
 package consul
 
-import libserf "github.com/hashicorp/consul/lib/serf"
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/hashicorp/consul/internal/resource"
+	"github.com/hashicorp/consul/internal/storage"
+	libserf "github.com/hashicorp/consul/lib/serf"
+	"github.com/hashicorp/consul/proto-public/pbresource"
+	pbtenancy "github.com/hashicorp/consul/proto-public/pbtenancy/v2beta1"
+	"github.com/hashicorp/consul/testrpc"
+)
 
 func updateSerfTags(s *Server, key, value string) {
 	libserf.UpdateTag(s.serfLAN, key, value)
@@ -13,4 +25,42 @@ func updateSerfTags(s *Server, key, value string) {
 	if s.serfWAN != nil {
 		libserf.UpdateTag(s.serfWAN, key, value)
 	}
+}
+
+func TestServer_InitTenancy(t *testing.T) {
+	t.Parallel()
+
+	_, conf := testServerConfig(t)
+	deps := newDefaultDeps(t, conf)
+	deps.Experiments = []string{"v2tenancy"}
+	deps.Registry = NewTypeRegistry()
+
+	s, err := newServerWithDeps(t, conf, deps)
+	require.NoError(t, err)
+
+	// first initTenancy call happens here
+	waitForLeaderEstablishment(t, s)
+	testrpc.WaitForLeader(t, s.RPC, "dc1")
+
+	nsID := &pbresource.ID{
+		Type:    pbtenancy.NamespaceType,
+		Tenancy: resource.DefaultPartitionedTenancy(),
+		Name:    resource.DefaultNamespaceName,
+	}
+
+	ns, err := s.resourceServiceServer.Backend.Read(context.Background(), storage.StrongConsistency, nsID)
+	require.NoError(t, err)
+	require.Equal(t, resource.DefaultNamespaceName, ns.Id.Name)
+
+	// explicitly call initiTenancy to verify we do not re-create namespace
+	err = s.initTenancy(context.Background(), s.resourceServiceServer.Backend)
+	require.NoError(t, err)
+
+	// read again
+	actual, err := s.resourceServiceServer.Backend.Read(context.Background(), storage.StrongConsistency, nsID)
+	require.NoError(t, err)
+
+	require.Equal(t, ns.Id.Uid, actual.Id.Uid)
+	require.Equal(t, ns.Generation, actual.Generation)
+	require.Equal(t, ns.Version, actual.Version)
 }
