+++ agent/consul/server_ce.go
@@ -6,7 +6,6 @@
 package consul
 
 import (
-	"context"
 	"fmt"
 	"strings"
 	"time"
@@ -21,7 +20,6 @@ import (
 	"github.com/hashicorp/consul/acl"
 	"github.com/hashicorp/consul/agent/consul/reporting"
 	"github.com/hashicorp/consul/agent/structs"
-	"github.com/hashicorp/consul/internal/storage"
 	"github.com/hashicorp/consul/lib"
 )
 
@@ -193,9 +191,4 @@ func (s *Server) updateReportingConfig(config ReloadableConfig) {
 func getEnterpriseReportingDeps(deps Deps) reporting.EntDeps {
 	// no-op
 	return reporting.EntDeps{}
-}
-
-func (s *Server) createDefaultPartition(ctx context.Context, b storage.Backend) error {
-	// no-op
-	return nil
 }
